Hotfix/mobile language switcher (#37)
* Merge develop into main with hot fixes for typescript shenanigans (#36)

* [CAN-18] Design: Stylise the localisation switch
Problem
Added styling for the switcher, along with flag svgs
Solution
Used the mobile nav as a template for the solution (thanks fab)
Note
A known issue is that once the dropdown for the switcher has been
expanded, then changing url/tab will not auto close it. Before
tackling the solutions I have found online I wanted to check if
the team is happy with the current styling and behaviour.

* [CAN-18] Design: Changing language switcher to onMouseEnter
Problem
This still doesnt solve the issue of an open dropdown menu being open
whilst clicking onto another link/tab, however it does make it more
unlikely for the dropdown menu to be open upon switching tabs.
Solution
Note

* [CAN-18] Toggle between mobile and desktop switcher designs
Problem
Depending on if the device has a sm width or higher, use custom desktop
language switcher, else use the select/option design which uses a
phone's native design.
Solution
combination of sm:visible invisible and sm:invisible visible to toggle
Note
This still does not solve the issue of the switcher being left open upon
switching tabs/urls

* [CAN-45] Adds on hover animation to desktop navbar links
Problem
Thought it would look a bit more alive to have active links with an
animation upon hovering.
Solution
Thanks to this stackoverflow solution:
https://stackoverflow.com/questions/70906977/tailwind-underline-hover-animation
Note
Very open to the styling being changed in the future, eg
colour/size/animation etc.

* [CAN-38] Fix: normalise club card sizes
Problem
Both the image size and the size of the card were varying across clubs,
the inconsistency did not look nice
Solution
Thanks to this stackoverflow:
https://stackoverflow.com/questions/77275607/how-to-make-images-same-size-next-image-component
i managed to get it to work.
Note
Feel free to change the sizes of the cards and the images. We need to be
wary that these are fixed and what will happen when content overflows.
Still need to test on multiple device sizes.

* [CAN-41] Fixes the language switcher remaining upon locale change
Problem
A long list of problems arose from attempting to solve this issue:
1) When clicking away from the switcher without choosing a language and
   subsequently loading a new locale route, the switcher component
   wouldnt close.
2) Then solving the above issue meant that the caret wouldn't switch on
   every click away from the switcher as well, meaning the toggle was no
   longer in sync
Solution
1) The solution to this problem was resolved thanks to this amazing
   stackoverflow:
   https://stackoverflow.com/questions/63086609/how-to-close-dropdown-on-outside-click-next-js
2) The solution to this problem was resolved based on this medium
   article (the former solution) as well as the linked repo:
   https://thivi.medium.com/react-development-94f95db41df6
   https://github.com/thivi/child-method-call-from-parent-react/tree/master
Note
Additionally, fixed some of the errors and warnings to do with the
incorrect react variable names which previously included hyphons.

* [CAN-41] Reformats mobile language switcher text
Problem
Solution
Note

* [CAN-41] Swaps us flag for uk flag, NEEDS STYLING
Problem
Solution
Note

* [CAN-42] Fixes switcher flag from US to UK flag
Problem
Had issues with the sizing of the UK flag that I imported
Solution
Used a fix size for both the german and UK flag
Note
Still need to fix all of the SVG in public

* [CAN-54] Center language switcher positioning
Problem
Switcher is at top of page, we want it center with rest of links.
Solution
Make the parent div flexbox so it can inherit the auto margins and
properties of it's parent flex container.
Note

* [CAN-41] Blocks use of language switcher whilst page is in transition
Problem
Block was only in use for mobile language switcher
Solution
Make sure to use the 'isPending' state for the whenever the dropdown
content is clicked, to avoid extreme cases where a user is able to spam
the switcher faster than the page can load.
Note
MAY NEED TO BE USER TESTED

* Hotfix/typescript for children (#35)

* HOTFIX - Typescript being a dick not working yet

* HOTFIX - typescript now happy but still a child

---------

Co-authored-by: marcaufderheyde <marc.auf-der-heyde@deutschebahn.com>

---------

Co-authored-by: Hector Jones <jjones.hector@gmail.com>
Co-authored-by: joneshector <89353359+joneshector@users.noreply.github.com>
Co-authored-by: Fabrizio <fabiecatinella@gmail.com>
Co-authored-by: marcaufderheyde <marc.auf-der-heyde@deutschebahn.com>

* [HOTFIX] Swaps US flag for UK flag on mobile
Problem
Fully changed from US to UK flag and also there was a conflict of the
label spilling out of the select/option component
Solution
Removed label as there is no need
Note

* [HOTFIX] Rounds the edges of mobile nav language switcher
Problem
Solution
Note

---------

Co-authored-by: Marc Auf der Heyde <marcaufderheyde@gmail.com>
Co-authored-by: Fabrizio <fabiecatinella@gmail.com>
Co-authored-by: marcaufderheyde <marc.auf-der-heyde@deutschebahn.com>