[CAN-63, CAN-99] Adds navbar as overlay onto map component Problem To align with fab's original designs, required layout.tsx override for the clubs page so that our header and footer default layout isnt apart of our map page. Solution Currently using route groups: https://nextjs.org/docs/app/building-your-application/routing/route-groups#opting-specific-segments-into-a-layout
Also required to add props to Navbar component, so that the zAxis can be
bumped up if we are showing the map page. As of right now this solution
leads to a deadend when switching from the Map View to the List of
clubs. Also there is a big overflow issue
Note